fix(MastheadLeftNav): set back button in menu to use button element
