+++ packages/react/src/components/carbon-components-react/UIShell/SideNavMenuSection.js
@@ -5,16 +5,13 @@
  * LICENSE file in the root directory of this source tree.
  */
 import ArrowRight20 from '@carbon/icons-react/es/arrow--right/20';
-import ChevronLeft20 from '@carbon/icons-react/es/chevron--left/20';
 import cx from 'classnames';
 import SideNavLink from '../../../internal/vendor/carbon-components-react/components/UIShell/SideNavLink';
-import SideNavMenuItem from '../../../internal/vendor/carbon-components-react/components/UIShell/SideNavMenuItem';
+import SideNavMenuBackButton from './SideNavMenuBackButton';
 import settings from 'carbon-components/es/globals/js/settings';
-import ddsSettings from '@carbon/ibmdotcom-utilities/es/utilities/settings/settings';
 import PropTypes from 'prop-types';
 import React, { useRef, useEffect } from 'react';
 
-const { stablePrefix } = ddsSettings;
 const { prefix } = settings;
 
 /**
@@ -68,34 +65,16 @@ const SideNavMenuSection = ({
     [customClassName]: !!customClassName,
   });
 
-  const handleBackButtonClick = event => {
-    event.preventDefault();
-    onBackClick();
-  };
-
-  const handleBackButtonKeyPress = event => {
-    if (event.key === 'Enter' || event.charCode === ' ') {
-      event.preventDefault();
-      onBackClick();
-    }
-  };
-
   return (
     <div className={className} id={rest.id} ref={menuSectionRef}>
-      {rest.backButtonText && (
+      {rest.isSubmenu && (
         <>
-          <SideNavMenuItem
+          <SideNavMenuBackButton
             ref={backButtonRef}
-            onClick={handleBackButtonClick}
-            onKeyPress={handleBackButtonKeyPress}
-            className={`${prefix}--masthead__side-nav--submemu-back`}
-            data-autoid={`${stablePrefix}--masthead-${rest.navType}-sidenav__l0-back`}
-            isbackbutton="true"
-            role="button"
-            tabIndex="0">
-            <ChevronLeft20 />
-            {rest.backButtonText}
-          </SideNavMenuItem>
+            onBackClick={onBackClick}
+            navType={rest.navType}
+            backButtonText={rest.backButtonText}
+          />
           {rest.titleUrl ? (
             <SideNavLink
               className={`${prefix}--masthead__side-nav--submemu-section-title`}
