+++ packages/styles/scss/components/masthead/_masthead-leftnav.scss
@@ -201,9 +201,24 @@
       top: 0;
       background-color: $ui-background;
       z-index: 1;
-      a {
+      button {
+        width: 100%;
+        border: none;
         cursor: pointer;
-        border-bottom: none;
+        outline: 2px solid transparent;
+        outline-offset: -2px;
+        position: relative;
+        display: flex;
+        align-items: center;
+        min-height: 2rem;
+        padding: 0 1rem;
+        background-color: $ui-background;
+
+        &:focus {
+          outline: 2px solid $focus;
+          outline-offset: -2px;
+        }
+
         > .#{$prefix}--side-nav__link-text {
           @include use-carbon-productive-tokens();
           @include carbon--type-style(body-short-02);
@@ -248,17 +263,20 @@
     overflow: hidden;
   }
 
-  a.#{$prefix}--side-nav__link {
-    height: carbon--mini-units(6);
-    @include carbon--type-style(body-short-02);
+  a,
+  button {
+    &.#{$prefix}--side-nav__link {
+      height: carbon--mini-units(6);
+      @include carbon--type-style(body-short-02);
 
-    > {
-      .#{$prefix}--side-nav__link-text {
-        @include carbon--type-style(body-short-02);
+      > {
+        .#{$prefix}--side-nav__link-text {
+          @include carbon--type-style(body-short-02);
 
-        color: $text-01;
-        width: 100%;
-        text-overflow: ellipsis;
+          color: $text-01;
+          width: 100%;
+          text-overflow: ellipsis;
+        }
       }
     }
   }
